Split code to individual components
